gcp_cloud_functions/vision_bigquery/requirements.txt
google-cloud-vision
google-cloud-bigquery